samplesite/requirements.txt
amqp==5.0.6
asgiref==3.4.1
astroid==2.8.5
beautifulsoup4==4.10.0
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.0
certifi==2021.10.8
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coverage==6.0.2
cycler==0.11.0
Django==3.2.7
django-bootstrap4==3.0.1
django-celery-results==2.2.0
django-cors-headers==3.10.0
django-elasticsearch-dsl==7.2.0
django-environ==0.8.1
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
djangorestframework==3.12.4
elasticsearch==7.15.1
elasticsearch-dsl==7.4.0
Faker==9.8.0
gunicorn==20.1.0
importlib-metadata==2.1.1
isort==5.10.1
kiwisolver==1.3.2
kombu==5.2.1
lazy-object-proxy==1.6.0
mccabe==0.6.1
Pillow==8.4.0
platformdirs==2.4.0
prompt-toolkit==3.0.22
psycopg2-binary==2.9.1
pylint==2.11.1
pyparsing==3.0.4
python-dateutil==2.8.2
pytz==2021.1
redis==3.5.3
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
transliterate==1.10.2
typed-ast==1.5.0
typing-extensions==3.10.0.2
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
zipp==3.6.0
